+++ src/HierarchicalPath.php
@@ -12,8 +12,9 @@
  */
 namespace League\Uri\Components;
 
+use Countable;
+use IteratorAggregate;
 use League\Uri\Components\Traits\PathInfo;
-use League\Uri\Interfaces\CollectionComponent;
 use League\Uri\Interfaces\PathComponent;
 use Traversable;
 
@@ -25,7 +26,7 @@ use Traversable;
  * @author     Ignace Nyamagana Butera <nyamsprod@gmail.com>
  * @since      1.0.0
  */
-class HierarchicalPath extends HierarchicalComponent implements PathComponent, CollectionComponent
+class HierarchicalPath extends HierarchicalComponent implements PathComponent, Countable, IteratorAggregate
 {
     use PathInfo;